requirement.txt
streamlit==1.30.1
torch==2.2.0
torchvision==0.15.2
numpy
pandas
matplotlib
pillow
